refactor(routing): migrate manage pages to app router
- Remove admin projects pages from _pages directory
- Migrate manage pages from _pages to app router structure
- Add Suspense boundaries and 'use client' directives
- Create manage layout with sidebar and providers
- Update SEOConfig to use client component
- Refactor resource hooks and utilities
- Move manage page widget to widgets directory
