--- _pages/[language]/admin/projects/milestones/index.tsx
@@ -1,48 +1,0 @@
-import { getAdminProjectLayout } from '@/layout/features/getProjectLayout';
-import { useSearchParams } from 'next/navigation';
-import { Skeleton } from '@/shared/ui/skeleton';
-import { useProjectMilestones } from '@/features/projects/hooks/milestone';
-import EmptyList from '@/components/Projects/ProjectList/EmptyList';
-import MilestoneItemView from '@/components/Milestones/MilestoneItemView';
-import { MilestonesProvider } from '@/contexts/Milestones';
-import { parseToString } from '@/shared/lib/helper';
-
-const ProjectMilestonesPage = () => {
-  const searchParams = useSearchParams();
-  const projectId = parseToString(searchParams?.get('id'));
-  const { data: milestones, isLoading } = useProjectMilestones(projectId);
-
-  return (
-    <div className="w-[750px] max-w-full mx-auto">
-      {isLoading ? (
-        <>
-          <Skeleton className="w-full h-[120px] mb-3" />
-          <Skeleton className="w-full h-[300px]" />
-        </>
-      ) : (
-        <div>
-          {projectId && Array.isArray(milestones) && milestones.length > 0 ? (
-            milestones
-              .sort((a, b) => {
-                return a.position - b.position;
-              })
-              .map((milestone, index) => {
-                return (
-                  <MilestoneItemView
-                    milestone={milestone}
-                    key={milestone.id}
-                    index={index}
-                  />
-                );
-              })
-          ) : (
-            <EmptyList />
-          )}
-        </div>
-      )}
-    </div>
-  );
-};
-ProjectMilestonesPage.getLayout = (page: React.ReactElement) =>
-  getAdminProjectLayout(<MilestonesProvider>{page}</MilestonesProvider>);
-export default ProjectMilestonesPage;
