--- _pages/[language]/admin/projects/notes/index.tsx
@@ -1,40 +1,0 @@
-import { useSearchParams } from 'next/navigation';
-import { ContentCard } from '@/features/projects';
-import { getAdminProjectLayout } from '@/layout/features/getProjectLayout';
-import { useProjectNotes } from '@/services/projects';
-import { parseToString } from '@/shared/lib/helper';
-
-const NotesPage = () => {
-  const searchParams = useSearchParams();
-  const projectId = parseToString(searchParams?.get('id'));
-
-  const { data: notes } = useProjectNotes(projectId);
-
-  if (!projectId) {
-    return <div>專案不存在</div>;
-  }
-
-  return (
-    <>
-      <ul className="px-4 bg-basic-white flex flex-col rounded-2xl">
-        {notes?.map((note) => (
-          <li
-            key={note.id}
-            className="py-6 border-b last:border-b-0 border-solid border-basic-200"
-          >
-            <ContentCard
-              type="note"
-              data={note}
-              className="p-3 transition-shadow hover:shadow-basic-200/40 hover:shadow-lg"
-              detailLink={`/admin/projects/notes/detail?id=${projectId}&noteId=${note.id}`}
-            />
-          </li>
-        ))}
-      </ul>
-    </>
-  );
-};
-
-NotesPage.getLayout = getAdminProjectLayout;
-
-export default NotesPage;
