--- _pages/[language]/admin/projects/outcomes/index.tsx
@@ -1,40 +1,0 @@
-import { useSearchParams } from "next/navigation";
-import { ContentCard } from "@/features/projects";
-import { getAdminProjectLayout } from "@/layout/features/getProjectLayout";
-import { useProjectOutcomes } from "@/features/projects/hooks/outcome";
-import { parseToString } from "@/shared/lib/helper";
-
-const OutcomesPage = () => {
-  const searchParams = useSearchParams();
-  const projectId = parseToString(searchParams?.get('id'));
-
-  const { data: outcomes } = useProjectOutcomes(projectId);
-
-  if (!projectId) {
-    return <div>專案不存在</div>;
-  }
-
-  return (
-    <>
-      <ul className="px-4 bg-basic-white flex flex-col rounded-2xl">
-        {outcomes?.map((outcome) => (
-          <li
-            key={outcome.id}
-            className="py-6 border-b last:border-b-0 border-solid border-basic-200"
-          >
-            <ContentCard
-              type="outcome"
-              data={outcome}
-              className="p-3 transition-shadow hover:shadow-basic-200/40 hover:shadow-lg"
-              detailLink={`/admin/projects/outcomes/detail?id=${projectId}&outcomeId=${outcome.id}`}
-            />
-          </li>
-        ))}
-      </ul>
-    </>
-  );
-};
-
-OutcomesPage.getLayout = getAdminProjectLayout;
-
-export default OutcomesPage;
