--- _pages/[language]/manage/groups/index.tsx
@@ -1,15 +1,0 @@
-import SEOConfig from "@/components/SEOConfig";
-import getManageLayout from "@/layout/features/getManageLayout";
-
-export default function GroupsPage() {
-  return (
-    <>
-      <SEOConfig title="揪團｜島島阿學" />
-      <div className="flex items-center justify-center min-h-screen">
-        <h1 className="text-2xl font-bold text-basic-400">揪團</h1>
-      </div>
-    </>
-  );
-}
-
-GroupsPage.getLayout = getManageLayout;
