--- _pages/[language]/manage/practices/index.tsx
@@ -1,15 +1,0 @@
-import SEOConfig from "@/components/SEOConfig";
-import getManageLayout from "@/layout/features/getManageLayout";
-
-export default function PracticesPage() {
-  return (
-    <>
-      <SEOConfig title="主題實踐｜島島阿學" />
-      <div className="flex items-center justify-center min-h-screen">
-        <h1 className="text-2xl font-bold text-basic-400">主題實踐</h1>
-      </div>
-    </>
-  );
-}
-
-PracticesPage.getLayout = getManageLayout;
