+++ app/[language]/(protected)/(default-layout)/manage/ideas/page.tsx
@@ -1,0 +1,17 @@
+import { Metadata } from 'next';
+import SEOConfig from '@/components/SEOConfig';
+
+export const metadata: Metadata = {
+  title: '想法｜島島阿學',
+};
+
+export default function IdeasPage() {
+  return (
+    <>
+      <SEOConfig title="想法｜島島阿學" />
+      <div className="flex min-h-screen items-center justify-center">
+        <h1 className="text-2xl font-bold text-basic-400">想法</h1>
+      </div>
+    </>
+  );
+}
