+++ app/[language]/(protected)/(default-layout)/manage/projects/milestones/page.tsx
@@ -1,0 +1,327 @@
+'use client';
+
+import { useMemo, useRef, useState } from 'react';
+import { usePathname } from 'next/navigation';
+import SEOConfig, { JsonLdType } from '@/components/SEOConfig';
+import { Skeleton } from '@/shared/ui/skeleton';
+import { AlignLeft } from 'lucide-react';
+import { Panel, Title, ProgressBar } from '@/components/Milestones/Shared';
+import { useProject } from '@/contexts/Project';
+import { MilestonesProvider } from '@/contexts/Milestones/index';
+import MilestoneCard, {
+  MilestoneFormRef,
+} from '@/components/Milestones/MilestoneCard';
+import DraggableMilestones from '@/components/Milestones/DraggableMilestones';
+import { isAfter, differenceInDays } from 'date-fns';
+import { DateRangePicker } from '@/shared/ui/date-picker';
+import { Button } from '@/shared/ui/button';
+import { ProjectMilestoneSchema } from '@/services/projects';
+import SwapRightIcon from '@/public/assets/icons/swap-right.svg';
+import { useMilestonesDateRange } from '@/features/projects';
+import {
+  useProjectMilestoneMutation,
+  useProjectMilestones,
+} from '@/features/projects/hooks/milestone';
+
+const SkeletonMilestones = () => {
+  return (
+    <>
+      <Skeleton className="mb-3 h-[120px] w-full" />
+      <Skeleton className="h-[300px] w-full" />
+    </>
+  );
+};
+
+interface MilestonesProgressProps {
+  milestones?: ProjectMilestoneSchema[];
+}
+
+const MilestonesProgress = ({ milestones = [] }: MilestonesProgressProps) => {
+  const progressValue = useMemo(() => {
+    if (!Array.isArray(milestones)) return 0;
+
+    const completedCount = milestones.filter((m) => m.isCompleted).length;
+    const allCount = milestones.length;
+
+    if (allCount === 0) return 0;
+    return Math.round((completedCount / allCount) * 100);
+  }, [milestones]);
+
+  const milestoneEndDate = useMemo(() => {
+    if (!Array.isArray(milestones)) return null;
+
+    return milestones.reduce<Date | null>((d, m) => {
+      const mEndDate = new Date(m.endDate || new Date());
+      return d && isAfter(d, mEndDate) ? d : mEndDate;
+    }, null);
+  }, [milestones]);
+
+  const daysRemaining = useMemo(() => {
+    if (!milestoneEndDate || isAfter(new Date(), milestoneEndDate)) return 0;
+
+    return differenceInDays(milestoneEndDate, new Date());
+  }, [milestoneEndDate]);
+
+  const remainingMilestonesCount = useMemo(() => {
+    if (!Array.isArray(milestones)) return 0;
+
+    return milestones.filter((m) => !m.isCompleted).length;
+  }, [milestones]);
+
+  const remainingTasksCount = useMemo(() => {
+    if (!Array.isArray(milestones)) return 0;
+
+    return milestones
+      .flatMap((milestone) => [...milestone.tasks])
+      .filter((task) => !task.isCompleted).length;
+  }, [milestones]);
+
+  return (
+    <Panel className="mb-6 flex flex-col gap-3 bg-white md:gap-5 md:py-6">
+      <Title title="學習里程碑進度" className="mb-0" />
+      {Array.isArray(milestones) && milestones.length > 0 ? (
+        <>
+          <ProgressBar progress={progressValue} />
+          <p className="font-sans text-sm text-basic-300 md:text-base">
+            {remainingMilestonesCount === 0 ? (
+              <span>恭喜你完成所有學習里程碑！</span>
+            ) : (
+              <span>
+                還剩 {daysRemaining} 天可以完成剩下的
+                {remainingTasksCount ? (
+                  <span> {remainingTasksCount} 個任務</span>
+                ) : (
+                  <span> {remainingMilestonesCount} 個里程碑</span>
+                )}
+              </span>
+            )}
+          </p>
+        </>
+      ) : (
+        <p className="font-sans text-sm text-basic-300 md:text-base">
+          趕快新增學習里程碑吧！
+        </p>
+      )}
+    </Panel>
+  );
+};
+
+enum FilterEnum {
+  All = 'all',
+  Completed = 'completed',
+  Incomplete = 'incomplete',
+}
+
+const filterItems = [
+  {
+    label: '全部',
+    value: FilterEnum.All,
+    fn: () => true,
+  },
+  {
+    label: '未完成',
+    value: FilterEnum.Incomplete,
+    fn: (milestone: ProjectMilestoneSchema) => !milestone.isCompleted,
+  },
+  {
+    label: '已完成',
+    value: FilterEnum.Completed,
+    fn: (milestone: ProjectMilestoneSchema) => milestone.isCompleted,
+  },
+];
+
+const useHandleShowForm = () => {
+  const [isCreating, setIsCreating] = useState(false);
+  const formRef = useRef<MilestoneFormRef>(null);
+
+  const handleOpen = () => {
+    setIsCreating(true);
+    formRef.current?.focus();
+  };
+
+  const handleClose = () => {
+    setIsCreating(false);
+  };
+
+  return { formRef, isCreating, handleOpen, handleClose };
+};
+
+const MilestonesContent = () => {
+  const { formRef, isCreating, handleOpen, handleClose } = useHandleShowForm();
+  const [filterType, setFilterType] = useState(FilterEnum.All);
+  const [isAscending, setIsAscending] = useState(true);
+  const { project } = useProject();
+
+  const {
+    data: milestones,
+    isLoading,
+    mutate,
+  } = useProjectMilestones(project.id);
+
+  const milestonesDateRange = useMilestonesDateRange(project.id);
+
+  const { createMutation, updateMutation } = useProjectMilestoneMutation({
+    projectId: project.id,
+    updateMilestoneCache: mutate,
+  });
+
+  const projectId = project.id;
+  const isMarathonProject = !!project.eventId;
+
+  const sortedMilestones = useMemo(() => {
+    if (!Array.isArray(milestones)) return [];
+
+    const filterFn =
+      filterItems.find((item) => item.value === filterType)?.fn ?? (() => true);
+
+    const sortedData = milestones.filter(filterFn);
+
+    return isAscending ? sortedData : [...sortedData].reverse();
+  }, [milestones, isAscending, filterType]);
+
+  const date = useMemo(() => {
+    return {
+      from: milestonesDateRange.startDate,
+      to: milestonesDateRange.endDate,
+    };
+  }, [milestonesDateRange]);
+
+  return (
+    <div>
+      {isLoading ? (
+        <SkeletonMilestones />
+      ) : (
+        <>
+          <MilestonesProgress milestones={milestones} />
+          <Panel className="bg-white">
+            <Title title="學習里程碑 *" />
+            {project && !isMarathonProject && (
+              <>
+                {!isMarathonProject && (
+                  <div className="flex flex-col justify-between gap-2 pb-2.5 md:flex-row md:items-center">
+                    <div className="flex items-center gap-2">
+                      <p>時間：</p>
+                      <DateRangePicker
+                        value={date}
+                        separator={
+                          <SwapRightIcon className="h-4 w-4 text-basic-black/25" />
+                        }
+                        className="-mx-3 p-2"
+                        disabled
+                      />
+                    </div>
+                    <Button
+                      variant="default"
+                      className="w-full md:w-auto"
+                      onClick={handleOpen}
+                    >
+                      新增學習里程碑
+                    </Button>
+                  </div>
+                )}
+                <div className="flex flex-col justify-between gap-2 pb-2.5 md:flex-row">
+                  <div>
+                    顯示：
+                    {filterItems.map((item) => (
+                      <Button
+                        key={item.value}
+                        variant={
+                          filterType === item.value ? 'default' : 'outline'
+                        }
+                        className="mr-2 rounded-lg px-2.5"
+                        onClick={() => setFilterType(item.value)}
+                      >
+                        {item.label}
+                      </Button>
+                    ))}
+                  </div>
+                  <Button
+                    variant="outline"
+                    className="group flex items-center gap-2 rounded-lg px-2.5"
+                    onClick={() => setIsAscending(!isAscending)}
+                  >
+                    <AlignLeft className="size-6 text-primary-base group-hover:text-current" />
+                    {isAscending ? '舊到新' : '新到舊'}
+                  </Button>
+                </div>
+              </>
+            )}
+            <div className="flex flex-col gap-3 transition-opacity">
+              {isCreating && (
+                <div className="flex flex-col gap-2 rounded-xl bg-basic-100 p-2.5">
+                  <MilestoneCard
+                    ref={formRef}
+                    minDate={milestonesDateRange.minDate}
+                    maxDate={milestonesDateRange.maxDate}
+                    projectId={projectId}
+                    disabledChangeDate={isMarathonProject}
+                    milestones={milestones}
+                    isEditable
+                    defaultEditing
+                    onCancel={handleClose}
+                    onCreate={async (request) => {
+                      await createMutation.trigger(request);
+                      handleClose();
+                    }}
+                  />
+                </div>
+              )}
+              {Array.isArray(sortedMilestones) && (
+                <DraggableMilestones
+                  milestones={sortedMilestones}
+                  projectId={projectId}
+                  minDate={milestonesDateRange.minDate}
+                  maxDate={milestonesDateRange.maxDate}
+                  isAscending={isAscending}
+                  onUpdate={updateMutation.trigger}
+                  onReorder={updateMutation.trigger}
+                  onRefreshData={mutate}
+                  isEditable
+                />
+              )}
+            </div>
+          </Panel>
+        </>
+      )}
+    </div>
+  );
+};
+
+export default function MilestonesPage() {
+  const pathname = usePathname();
+  const jsonLd = useMemo<JsonLdType>(
+    () => ({
+      '@context': 'https://schema.org',
+      '@graph': [
+        {
+          '@context': 'https://schema.org',
+          '@type': 'WebSite',
+          url: 'https://www.daoedu.tw',
+          potentialAction: {
+            '@type': 'SearchAction',
+            'query-input': 'required name=q',
+            target: 'https://www.daoedu.tw/search?q={q}',
+          },
+        },
+        {
+          '@context': 'https://schema.org',
+          '@type': 'Organization',
+          url: 'https://www.daoedu.tw',
+          logo: 'https://www.daoedu.tw/assets/brand/favicon-112.png',
+        },
+      ],
+    }),
+    [pathname]
+  );
+
+  return (
+    <MilestonesProvider>
+      <SEOConfig
+        title="島島盃 - 2025 春季學習馬拉松｜多元學習資源平台｜島島阿學"
+        description="「島島阿學」盼能透過建立多元的學習資源網絡，讓自主學習者能找到合適的成長方法，進一步成為自己想成為的人，從中培養共好精神。目前正積極打造「可共編的學習資源平台」。"
+        jsonLd={jsonLd}
+      />
+      <MilestonesContent />
+    </MilestonesProvider>
+  );
+}
