+++ app/[language]/(protected)/(default-layout)/manage/projects/reviews/page.tsx
@@ -1,0 +1,150 @@
+'use client';
+
+import toast from 'react-hot-toast';
+import { useMemo, useState, Suspense } from 'react';
+import { useSearchParams } from 'next/navigation';
+import { Button } from '@/shared/ui/button';
+import ReviewCard from '@/features/projects/components/ReviewCard';
+import {
+  ReviewCreateModal,
+  ReviewDeleteModal,
+  ReviewUpdateModal,
+} from '@/features/projects';
+import SEOConfig from '@/components/SEOConfig';
+import { useProject } from '@/services/projects';
+import {
+  useProjectReview,
+  useProjectReviewList,
+} from '@/features/projects/hooks/review';
+import marathonConfig from '@/constants/marathon';
+import { parseToString } from '@/shared/lib/helper';
+
+enum ModalTypeEnum {
+  Create,
+  Update,
+  Delete,
+}
+
+function ReviewsPageContent() {
+  const searchParams = useSearchParams();
+  const projectId = parseToString(searchParams?.get('id'));
+  const [modalType, setModalType] = useState<ModalTypeEnum | null>(null);
+  const [reviewId, setReviewId] = useState<number | undefined>(undefined);
+  const { data: project } = useProject(projectId);
+
+  const { data: reviews } = useProjectReviewList(projectId);
+
+  const { data: detail } = useProjectReview({
+    projectId,
+    reviewId,
+  });
+
+  const SEOData = useMemo(
+    () => ({
+      title: `${project?.title} 覆盤｜島島阿學`,
+      description:
+        project?.description?.substring(0, 150) ||
+        '「島島阿學」盼能透過建立多元的學習資源網絡，讓自主學習者能找到合適的成長方法，進一步成為自己想成為的人，從中培養共好精神。目前正積極打造「可共編的學習資源平台」。',
+      keywords: '島島阿學',
+      author: '島島阿學',
+      copyright: '島島阿學',
+      imgLink: 'https://www.daoedu.tw/assets/brand/horizontal-primary-logo.svg',
+      link: `${process.env.PROD_URL}/manage/projects/reviews?id=${projectId}`,
+    }),
+    [project?.title, project?.description, projectId]
+  );
+
+  if (!projectId) {
+    return <div>專案不存在</div>;
+  }
+
+  return (
+    <>
+      <SEOConfig {...SEOData} />
+      <div className="body-md mb-6 flex items-end justify-between sm:items-center">
+        <div className="flex flex-col items-start gap-1 sm:flex-row sm:items-center">
+          <div className="text-basic-500">
+            覆盤（{marathonConfig.getWeekNumber().toString().padStart(2, '0')}{' '}
+            週/22週）
+          </div>
+        </div>
+        <Button
+          variant="default"
+          onClick={() => setModalType(ModalTypeEnum.Create)}
+        >
+          新增覆盤
+        </Button>
+      </div>
+
+      <ul className="flex flex-col gap-6">
+        {Array.isArray(reviews) &&
+          reviews.map((review) => (
+            <li key={review.id}>
+              <ReviewCard
+                data={review}
+                detailLink={`/manage/projects/reviews/detail?id=${projectId}&reviewId=${review.id}`}
+                onEditClick={() => {
+                  setReviewId(review.id);
+                  setModalType(ModalTypeEnum.Update);
+                }}
+                onDeleteClick={() => {
+                  setReviewId(review.id);
+                  setModalType(ModalTypeEnum.Delete);
+                }}
+              />
+            </li>
+          ))}
+      </ul>
+
+      {project && (
+        <ReviewCreateModal
+          projectId={projectId}
+          projectTitle={project.title}
+          isOpen={modalType === ModalTypeEnum.Create}
+          onClose={() => setModalType(null)}
+          onSuccess={() => {
+            toast.success('新增成功');
+            setModalType(null);
+          }}
+        />
+      )}
+
+      {detail && reviewId && project && (
+        <ReviewUpdateModal
+          projectId={projectId}
+          projectTitle={project.title}
+          reviewId={reviewId}
+          isOpen={modalType === ModalTypeEnum.Update}
+          onClose={() => setModalType(null)}
+          onSuccess={() => {
+            toast.success('更新成功');
+            setModalType(null);
+            setReviewId(undefined);
+          }}
+        />
+      )}
+
+      {reviewId && (
+        <ReviewDeleteModal
+          projectId={projectId}
+          reviewId={reviewId}
+          isOpen={modalType === ModalTypeEnum.Delete}
+          onClose={() => setModalType(null)}
+          onSuccess={() => {
+            toast.success('刪除成功');
+            setModalType(null);
+            setReviewId(undefined);
+          }}
+        />
+      )}
+    </>
+  );
+}
+
+export default function ReviewsPage() {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <ReviewsPageContent />
+    </Suspense>
+  );
+}
