+++ entities/resource/lib/get-resource-data.ts
@@ -22,7 +22,9 @@ export const getResourceListData = async (
 ) => {
   return getSwrKeyWithResponse('/api/v1/resources/', {
     params: {
-      query: searchParams,
+      query: {
+        cursor: searchParams?.cursor,
+      },
     },
   });
 };
