+++ widgets/manage/index.ts
@@ -1,0 +1,2 @@
+export { ManagePageWidget } from './ui/manage-page-widget';
+
